requirements.txt
dnspython~=2.6.1
idna~=2.10
httpcore~=1.0.5
httpx~=0.27.0
cryptography~=42.0.8
anyio~=4.3.0
h11~=0.14.0
pip~=24.0
attrs~=23.2.0
wheel~=0.42.0
contrib~=0.3.0
cloudbiolinux~=0.1
pillow~=10.3.0
tornado~=6.1
Jinja2~=3.1.3
filelock~=3.13.1
base~=0.0.0
typing_extensions~=4.9.0
rich~=13.7.1
Pygments~=2.17.2
PyYAML~=6.0.1
sniffio~=1.3.0
click~=8.1.7
certifi~=2023.11.17
mdurl~=0.1.2
typer~=0.12.3
shellingham~=1.5.4
ipython~=8.19.0
MarkupSafe~=2.1.5
filters~=1.3.2
orjson~=3.10.5
uvloop~=0.19.0
fastapi~=0.111.0
starlette~=0.37.2
email_validator~=2.2.0
pydantic~=2.5.3
pydantic_core~=2.14.6
requests~=2.31.0
websockets~=12.0
ujson~=5.10.0
uvicorn~=0.30.1
httptools~=0.6.1
gunicorn~=21.2.0
watchfiles~=0.22.0
h2~=3.2.0
Deprecated~=1.2.14
python-multipart~=0.0.9
itsdangerous~=2.1.2
state~=0.1.1.dev0
SQLAlchemy~=2.0.31
python-dateutil~=2.8.2
fastapi
sqlalchemy
uvicorn
passlib[bcrypt]
python-jose
python-dotenv
